Reorganize xarray files
