+++ setup.cfg
@@ -52,7 +52,7 @@ examples = matplotlib
 console_scripts =
     tiledb-cf = tiledb.cf:cli
 xarray.backends =
-    tiledb = tiledb.cf.xarray_engine.default_backend_engine:TileDBXarrayBackendEntrypoint
+    tiledb = tiledb.cf.xarray_engine.engine:TileDBXarrayBackendEntrypoint
 
 
 [flake8]
